requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Redis
redis==5.2.1

# Validation & settings
pydantic==2.10.4
pydantic-settings==2.7.1

# XML parsing
lxml==5.3.0

# HTTP client
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0

# Google Calendar OAuth2
google-auth-oauthlib==1.2.1
google-api-python-client==2.155.0
cryptography==44.0.0

# Google Gemini (Agent 4 — insight generation)
google-genai==0.8.0

# Background processing
celery==5.4.0
kombu==5.4.2

# Statistical analysis (Agent 3 — pattern discovery)
pandas==2.2.3
numpy==2.2.0
scipy==1.14.1

# Utilities
python-dateutil==2.9.0.post0
